Correction page panier (#42)
* fix: boutton ajouter panier

* fix: initialisation par défaut à 1 pour l'ajout d'un produit

* fix: ajout de la fonctionnalité "payer"

---------

+++ frontend/pages/product/[id].js
@@ -6,14 +6,32 @@ export default function ProductDetail() {
   const { id } = router.query;
   const [productByID, setProductByID] = useState(null);
   const [category, setCategory] = useState(null);
+  const [cart, setCart] = useState([]);
   const [performanceByProduct, setPerformanceByProduct] = useState(null);
   const [photoByProduct, setPhotoByProduct] = useState(null);
-
-
+  
+  useEffect(() => {
+      const storedCart = JSON.parse(localStorage.getItem("cart")) || [];
+      setCart(storedCart);
+    }, []);
+  
   const handleAddToCart = () => {
     const storedCart = JSON.parse(localStorage.getItem("cart")) || [];
-    const updatedCart = [...storedCart, productByID];
+    const existingProduct = storedCart.find((item) => item.id_product === productByID.id_product);
+
+  if (existingProduct) {
+    const updatedCart = storedCart.map((item) =>
+      item.id_product === productByID.id_product
+        ? { ...item, quantity: item.quantity + 1 }
+        : item
+    );
+    setCart(updatedCart);
     localStorage.setItem("cart", JSON.stringify(updatedCart));
+  } else {
+    const updatedCart = [...storedCart, { ...productByID, quantity: 1 }];
+    setCart(updatedCart);
+    localStorage.setItem("cart", JSON.stringify(updatedCart));
+  }
     alert(`${productByID.name} a été ajouté au panier !`);
   };
